writing-practice/requirements.txt
# Core dependencies
streamlit>=1.31.0,<2.0.0
numpy>=1.24.3,<2.0.0

# Image processing
Pillow>=10.1.0,<11.0.0

# API and HTTP
requests>=2.31.0,<3.0.0
groq>=0.4.0,<1.0.0

# Drawing canvas
streamlit_drawable_canvas>=0.9.3,<1.0.0

# Environment and configuration
python-dotenv>=1.0.0,<2.0.0

# Development tools
black>=24.2.0  # Code formatting
isort>=5.13.0  # Import sorting
mypy>=1.8.0    # Type checking
pytest>=8.0.0  # Testing
